requirements.txt
cffi==1.15.1
cryptography==40.0.2
customtkinter==5.1.3
darkdetect==0.8.0
Pillow==9.5.0
pycparser==2.21
pygame==2.4.0
PyMySQL==1.0.3
tkinter-tooltip==2.1.0
